Restructured cosmic
coffee script is working again, still having difficulties with the json tree that is being sent to the coffee script. I am trying to figure out one that can filter out characters, but it looks like I can only do that as a separate flat array instead of embedding it into the tree as mongoid's queries does not support that kind of filtering. I will have to use a scope to grab all characters and include references to where they should belong, I will have to revive the object.coordinates macro.

Signed-off-by: Mark Poon <markpoon@me.com>